Move helper functions to a new file
